Type fix

+++ utils.py
@@ -92,7 +92,7 @@ def postprocess_segmap_tensor_to_pil_img(segmap_tensor, apply_color=True, binari
     processed_segmap = foreground * damage_class
     if binarize:
         processed_segmap[processed_segmap > 1] = 1
-    r = Image.fromarray(processed_segmap)
+    r = Image.fromarray(processed_segmap.astype(np.uint8))
     if apply_color:
         r.putpalette(colors)
         return r
